+++ KeireCore/Include/KeireInternal/UiInputInternal.h
@@ -5,4 +5,6 @@
 namespace Keire::Detail
 {
     [[nodiscard]] bool UiBackendKeyDown(UiKey key) noexcept;
+    [[nodiscard]] bool UiBackendKeyPressed(UiKey key) noexcept;
+    [[nodiscard]] std::string UiBackendTextInput();
 } // namespace Keire::Detail